% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Semanticator.R, R/compare_words.R
\name{compare_words}
\alias{compare_words}
\title{Count matching sentiment words}
\usage{
compare_words(word_vector, sentiment_words)

compare_words(word_vector, sentiment_words)
}
\arguments{
\item{word_vector}{Vector of words from the document}

\item{sentiment_words}{Vector of positive or negative words}
}
\value{
Number of matches found

Number of matches found
}
\description{
Count matching sentiment words

Count matching sentiment words
}
