% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Semanticator.R, R/preprocess_text.R
\name{preprocess_text}
\alias{preprocess_text}
\title{Preprocess text from a .txt file}
\usage{
preprocess_text(filepath)

preprocess_text(filepath)
}
\arguments{
\item{filepath}{Absolute path to the text file}
}
\value{
A vector of lowercase, punctuation-free words

A vector of lowercase, punctuation-free words
}
\description{
Preprocess text from a .txt file

Preprocess text from a .txt file
}
