% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Semanticator.R, R/sentiment_count.R
\name{sentiment_count}
\alias{sentiment_count}
\title{Count positive and negative sentiment words}
\usage{
sentiment_count(word_vector, positive_words, negative_words)

sentiment_count(word_vector, positive_words, negative_words)
}
\arguments{
\item{word_vector}{Vector of words from the document}

\item{positive_words}{Vector of positive words}

\item{negative_words}{Vector of negative words}
}
\value{
A list with counts: positive and negative

A list with counts: positive and negative
}
\description{
Count positive and negative sentiment words

Count positive and negative sentiment words
}
